+++ ql/time/date.hpp
@@ -41,6 +41,8 @@
 #include <utility>
 #include <functional>
 
+#include <boost/cstdint.hpp>
+
 namespace QuantLib {
 
     //! Day number
@@ -124,7 +126,7 @@ namespace QuantLib {
         //! Default constructor returning a null date.
         Date();
         //! Constructor taking a serial number as given by Applix or Excel.
-        explicit Date(BigInteger serialNumber);
+        explicit Date(int_fast32_t serialNumber);
         //! More traditional constructor.
         Date(Day d, Month m, Year y);
 
@@ -146,7 +148,7 @@ namespace QuantLib {
         Day dayOfYear() const;
         Month month() const;
         Year year() const;
-        BigInteger serialNumber() const;
+        int_fast32_t serialNumber() const;
 
 #ifdef QL_HIGH_RESOLUTION_DATE
         Hour hours() const;
@@ -165,11 +167,11 @@ namespace QuantLib {
         //! \name date algebra
         //@{
         //! increments date by the given number of days
-        Date& operator+=(BigInteger days);
+        Date& operator+=(int_fast32_t days);
         //! increments date by the given period
         Date& operator+=(const Period&);
         //! decrement date by the given number of days
-        Date& operator-=(BigInteger days);
+        Date& operator-=(int_fast32_t days);
         //! decrements date by the given period
         Date& operator-=(const Period&);
         //! 1-day pre-increment
@@ -181,11 +183,11 @@ namespace QuantLib {
         //! 1-day post-decrement
         Date operator--(int );
         //! returns a new date incremented by the given number of days
-        Date operator+(BigInteger days) const;
+        Date operator+(int_fast32_t days) const;
         //! returns a new date incremented by the given period
         Date operator+(const Period&) const;
         //! returns a new date decremented by the given number of days
-        Date operator-(BigInteger days) const;
+        Date operator-(int_fast32_t days) const;
         //! returns a new date decremented by the given period
         Date operator-(const Period&) const;
         //@}
@@ -236,25 +238,25 @@ namespace QuantLib {
         //@}
 
       private:
-        static BigInteger minimumSerialNumber();
-        static BigInteger maximumSerialNumber();
-        static void checkSerialNumber(BigInteger serialNumber);
+        static int_fast32_t minimumSerialNumber();
+        static int_fast32_t maximumSerialNumber();
+        static void checkSerialNumber(int_fast32_t serialNumber);
 
 #ifdef QL_HIGH_RESOLUTION_DATE
         boost::posix_time::ptime dateTime_;
 #else
-        BigInteger serialNumber_;
+        int_fast32_t serialNumber_;
         static Date advance(const Date& d, Integer units, TimeUnit);
         static Integer monthLength(Month m, bool leapYear);
         static Integer monthOffset(Month m, bool leapYear);
-        static BigInteger yearOffset(Year y);
+        static int_fast32_t yearOffset(Year y);
 #endif
     };
 
     /*! \relates Date
         \brief Difference in days between dates
     */
-    BigInteger operator-(const Date&, const Date&);
+    int_fast32_t operator-(const Date&, const Date&);
     /*! \relates Date
         \brief Difference in days (including fraction of days) between dates
     */
@@ -366,15 +368,15 @@ namespace QuantLib {
         return serialNumber_ - yearOffset(year());
     }
 
-    inline BigInteger Date::serialNumber() const {
+    inline int_fast32_t Date::serialNumber() const {
         return serialNumber_;
     }
 
-    inline Date Date::operator+(BigInteger days) const {
+    inline Date Date::operator+(int_fast32_t days) const {
         return Date(serialNumber_+days);
     }
 
-    inline Date Date::operator-(BigInteger days) const {
+    inline Date Date::operator-(int_fast32_t days) const {
         return Date(serialNumber_-days);
     }
 
@@ -396,7 +398,7 @@ namespace QuantLib {
        return (d.dayOfMonth() == monthLength(d.month(), isLeap(d.year())));
     }
 
-    inline BigInteger operator-(const Date& d1, const Date& d2) {
+    inline int_fast32_t operator-(const Date& d1, const Date& d2) {
         return d1.serialNumber()-d2.serialNumber();
     }
