+++ ql/time/daycounters/business252.hpp
@@ -40,7 +40,7 @@ namespace QuantLib {
             Calendar calendar_;
           public:
             std::string name() const;
-            BigInteger dayCount(const Date& d1,
+            int_fast32_t dayCount(const Date& d1,
                                 const Date& d2) const;
             Time yearFraction(const Date& d1,
                               const Date& d2,
