+++ ql/time/daycounters/thirty360.hpp
@@ -60,7 +60,7 @@ namespace QuantLib {
         class US_Impl : public DayCounter::Impl {
           public:
             std::string name() const { return std::string("30/360 (Bond Basis)");}
-            BigInteger dayCount(const Date& d1,
+            int_fast32_t dayCount(const Date& d1,
                                 const Date& d2) const;
             Time yearFraction(const Date& d1,
                               const Date& d2,
@@ -71,7 +71,7 @@ namespace QuantLib {
         class EU_Impl : public DayCounter::Impl {
           public:
             std::string name() const { return std::string("30E/360 (Eurobond Basis)");}
-            BigInteger dayCount(const Date& d1,
+            int_fast32_t dayCount(const Date& d1,
                                 const Date& d2) const;
             Time yearFraction(const Date& d1,
                               const Date& d2,
@@ -82,7 +82,7 @@ namespace QuantLib {
         class IT_Impl : public DayCounter::Impl {
           public:
             std::string name() const { return std::string("30/360 (Italian)");}
-            BigInteger dayCount(const Date& d1, const Date& d2) const;
+            int_fast32_t dayCount(const Date& d1, const Date& d2) const;
             Time yearFraction(const Date& d1,
                               const Date& d2,
                               const Date&,
